# container/docs/swagger.yaml
definitions:
...
  github_com_vanpelt_catnip_internal_services.ServiceInfo:
    properties:
      command:
        type: string
      health:
        type: string
      last_seen:
        type: string
      pid:
        type: integer
      port:
        type: integer
      service_type:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing container PTY sessions and services
  title: Catnip Container API
  version: "1.0"
paths:
...
  /v1/ports/{port}:
    get:
      consumes:
      - application/json
      description: Returns detailed information about a specific port if it exists
      parameters:
      - description: Port number
        in: path
        name: port
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Port information
          schema:
            $ref: '#/definitions/github_com_vanpelt_catnip_internal_services.ServiceInfo'
        "404":
          description: Port not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get port information
      tags:
      - ports
...
schemes:
- http
- ws
swagger: "2.0"